rework disable gif function.
Previous one doesn't work on many website, so replace it by a gif decoder Thanks to https://github.com/matt-way/gifuct-js to extract first frame of the gif.

Remove the button to enable/disable on each gif as the button break quite all website.

The best way to avoid animated gif is to contact the website webmaster to explain why he need to replace it.

TODO:
- test IE11 extension
- see why data.set seems to doesn't work on firefox, but need to debug it